Resolved #542:Fallback to web.config (or app.config) if ISettingStore is not implemented.
